requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
anywidget==0.9.21
appdirs==1.4.4
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
babel==2.18.0
backports.zstd==1.3.0
beautifulsoup4==4.14.3
boto3==1.40.61
botocore==1.40.61
Brotli==1.2.0
cached-property==1.5.2
certifi==2026.2.25
charset-normalizer==3.4.5
chatlas==0.15.2
chevron==0.14.0
click==8.2.1
comm==0.2.3
commonmark==0.9.1
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.4.4
exceptiongroup==1.3.1
executing==2.2.1
faicons==0.2.2
filelock==3.25.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.9.0
great-tables==0.21.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.1
hpack==4.1.0
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
hyperframe==6.1.0
idna==3.11
ijson==3.5.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inspect_ai==0.3.188
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
linkify-it-py==2.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
munkres==1.1.4
narwhals==2.17.0
nest-asyncio2==1.7.2
numpy==2.4.2
openai==2.26.0
orjson==3.11.7
packaging==26.0
pandas==3.0.0
parso==0.8.6
pathlib_abc==0.5.2
patsy==1.0.2
pexpect==4.9.0
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.2
playwright=1.58.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
psygnal==0.15.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
querychat==0.5.1
questionary==2.1.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
s3fs==2025.9.0
s3transfer==0.14.0
scipy==1.17.1
seaborn==0.13.2
semver==3.0.4
setuptools==82.0.0
shellingham==1.5.4
shiny==1.5.1
shinychat==0.2.9
shinywidgets==0.7.1
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack_data==0.6.3
starlette==0.52.1
statsmodels==0.14.6
tenacity==9.1.4
textual==8.0.2
tiktoken==0.12.0
tokenizers==0.22.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.0
typer-slim==0.24.0
typing_extensions==4.15.0
typing-inspection==0.4.2
uc-micro-py==1.0.3
unicodedata2==17.0.1
universal_pathlib==0.3.10
urllib3==2.6.3
uvicorn==0.41.0
vega_datasets==0.9.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wheel==0.46.3
widgetsnbextension==4.0.15
wrapt==1.17.3
yarl==1.23.0
zipfile-zstd==0.0.4
zipp==3.23.0
zstandard==0.25.0
